requirements.txt
numpy==1.19.5
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
tensorflow-gpu==2.5.0
